move out block scoping tests
